moved stacking code into separate module
