+++ banzai/tests/test_main.py
@@ -8,7 +8,7 @@ from dramatiq.brokers.stub import StubBroker
 from dramatiq.actor import Actor
 from dramatiq.message import Message
 
-from banzai.main import schedule_stacking_checks, schedule_stack, should_retry_schedule_stack
+from banzai.tasks import schedule_stacking_checks, schedule_stack, should_retry_schedule_stack
 from banzai.settings import CALIBRATION_STACK_DELAYS
 from banzai.utils import date_utils
 from banzai.context import Context
@@ -77,10 +77,10 @@ def stub_worker(stub_broker):
 
 
 class TestMain():
-    @mock.patch('banzai.main.schedule_stack.send_with_options')
-    @mock.patch('banzai.main.dbs.get_instruments_at_site')
-    @mock.patch('banzai.main.lake_utils.get_calibration_blocks_for_time_range')
-    @mock.patch('banzai.main.lake_utils.filter_calibration_blocks_for_type')
+    @mock.patch('banzai.tasks.schedule_stack.send_with_options')
+    @mock.patch('banzai.tasks.dbs.get_instruments_at_site')
+    @mock.patch('banzai.tasks.lake_utils.get_calibration_blocks_for_time_range')
+    @mock.patch('banzai.tasks.lake_utils.filter_calibration_blocks_for_type')
     def test_schedule_stacking_checks_queues_task_no_delay(self, mock_filter_blocks, mock_get_blocks,     mock_get_instruments, mock_schedule_stack, stub_broker):
         actor = Actor(fn=foo, broker=stub_broker, actor_name='', queue_name='schedule_stack', priority=1, options={})
         mock_get_instruments.return_value = [FakeInstrument(site='coj', camera='2m0-SciCam-Spectral',
@@ -91,12 +91,12 @@ class TestMain():
         runtime_context = Context(runtime_context_json)
         schedule_stacking_checks(runtime_context)
         mock_schedule_stack.assert_called_with(args=(runtime_context._asdict(), mock_filter_blocks.return_value,    'BIAS', 'coj', '2m0-SciCam-Spectral', 'clma', '2m0a'), delay=0, kwargs={'process_any_images': False},  on_failure=mock.ANY)
-        assert stub_broker.queues['schedule_stack.DQ'].qsize() == 3
+        assert stub_broker.queues['schedule_stack.DQ'].qsize() == 1
 
-    @mock.patch('banzai.main.schedule_stack.send_with_options')
-    @mock.patch('banzai.main.dbs.get_instruments_at_site')
-    @mock.patch('banzai.main.lake_utils.get_calibration_blocks_for_time_range')
-    @mock.patch('banzai.main.lake_utils.filter_calibration_blocks_for_type')
+    @mock.patch('banzai.tasks.schedule_stack.send_with_options')
+    @mock.patch('banzai.tasks.dbs.get_instruments_at_site')
+    @mock.patch('banzai.tasks.lake_utils.get_calibration_blocks_for_time_range')
+    @mock.patch('banzai.tasks.lake_utils.filter_calibration_blocks_for_type')
     def test_schedule_stacking_checks_queues_task_with_delay(self, mock_filter_blocks, mock_get_blocks,   mock_get_instruments, mock_schedule_stack, stub_broker):
         actor = Actor(fn=foo, broker=stub_broker, actor_name='', queue_name='schedule_stack', priority=1, options={})
         mock_get_instruments.return_value = [FakeInstrument(site='coj', camera='2m0-SciCam-Spectral',
@@ -111,11 +111,11 @@ class TestMain():
                                                      'BIAS', 'coj', '2m0-SciCam-Spectral', 'clma', '2m0a'),
                                                delay=(60000+CALIBRATION_STACK_DELAYS['BIAS']),
                                                kwargs={'process_any_images': False}, on_failure=mock.ANY)
-        assert stub_broker.queues['schedule_stack.DQ'].qsize() == 3
+        assert stub_broker.queues['schedule_stack.DQ'].qsize() == 1
 
-    @mock.patch('banzai.main.process_master_maker')
-    @mock.patch('banzai.main.dbs.get_individual_calibration_images')
-    @mock.patch('banzai.main.dbs.query_for_instrument')
+    @mock.patch('banzai.calibrations.process_master_maker')
+    @mock.patch('banzai.tasks.dbs.get_individual_calibration_images')
+    @mock.patch('banzai.tasks.dbs.query_for_instrument')
     def test_schedule_stack(self, mock_query_for_inst, mock_get_calibration_images, mock_process_master_maker):
         fake_inst = FakeInstrument(site='coj', camera='2m0-SciCam-Spectral', enclosure='clma', telescope='2m0a')
         mock_query_for_inst.return_value = fake_inst
@@ -129,9 +129,9 @@ class TestMain():
                                                      expected_runtime_context.min_date,
                                                      expected_runtime_context.max_date)
 
-    @mock.patch('banzai.main.process_master_maker')
-    @mock.patch('banzai.main.dbs.get_individual_calibration_images')
-    @mock.patch('banzai.main.dbs.query_for_instrument')
+    @mock.patch('banzai.calibrations.process_master_maker')
+    @mock.patch('banzai.tasks.dbs.get_individual_calibration_images')
+    @mock.patch('banzai.tasks.dbs.query_for_instrument')
     def test_schedule_stack_not_enough_images(self, mock_query_for_inst, mock_get_calibration_images, mock_process_master_maker):
         fake_inst = FakeInstrument(site='coj', camera='2m0-SciCam-Spectral', enclosure='clma', telescope='2m0a')
         mock_query_for_inst.return_value = fake_inst
@@ -140,14 +140,14 @@ class TestMain():
             schedule_stack(runtime_context_json, [fake_blocks_response_json['results'][0]], 'BIAS', 'coj',
                            '2m0-SciCam-Spectral', 'clma', '2m0a')
 
-    @mock.patch('banzai.main.schedule_stack')
+    @mock.patch('banzai.tasks.schedule_stack')
     def test_should_retry_schedule_stack_one_retry(self, mock_schedule_stack):
         test_message = Message(options={'retries': 1}, args=('test_arg1', 'test_arg2'),
                                queue_name='schedule_stack', actor_name='', kwargs={})
         return_value = should_retry_schedule_stack(test_message._asdict(), {'type': Exception, 'message': 'test'})
         assert not return_value
 
-    @mock.patch('banzai.main.schedule_stack')
+    @mock.patch('banzai.tasks.schedule_stack')
     def test_should_retry_schedule_stack_two_retries(self, mock_schedule_stack):
         test_message = Message(options={'retries': 2}, args=('test_arg1', 'test_arg2'),
                                queue_name='schedule_stack', actor_name='', kwargs={})
