+++ setup.cfg
@@ -88,8 +88,8 @@ version = 0.21.0
 banzai = banzai.main:main
 banzai_reduce_individual_frame = banzai.main:reduce_single_frame
 banzai_reduce_directory = banzai.main:reduce_directory
-banzai_process_master_maker = banzai.main:process_master_maker
-banzai_stack_calibrations = banzai.main:schedule_calibration_stacking
+banzai_stack_calibrations = banzai.main:stack_calibrations
+banzai_e2e_stack_calibrations = banzai.main:e2e_stack_calibrations
 banzai_automate_stack_calibrations = banzai.main:start_schedule_calibration_stacking
 banzai_run_realtime_pipeline = banzai.main:run_realtime_pipeline
 banzai_mark_frame_as_good = banzai.main:mark_frame_as_good
